core: add add/remove edge methods to Passage
These are protected methods used to trigger updating of layers
even for Edges which aren't included in a specific layer, or
for any object in general.
Note that calling Passage._add_edge is done in Node.add(), and not in
Edge.__init__; this is because usually updating is based upon the
connection of the Node with others, and these variables aren't updated
when the Edge is created.

Signed-off-by: Amit Beka <amit.beka@mail.huji.ac.il>