core: always attach Edges to a layer
Updating the Passage/Layer objects of changed edges was only committed
if both Nodes of the Edge were of the same Layer. In practice, Edges
from one Layer to another are part of the subgraph of the parent Edge,
so treat them like ones and call the updaing of the Layer.

Signed-off-by: Amit Beka <amit.beka@mail.huji.ac.il>