scripts: add postag_passages.py
usage: postag_passages.py <pickle> <dir-with-text-files>
for each passage in the pickle, gets the corresponding text file called
pid.tagged where pid is the passage id, and puts the correct postag in
the Terminal extra dict under 'postags'. In the end saves all the
pickles to the same file name with '.tags' suffix.

Signed-off-by: Amit Beka <amit.beka@mail.huji.ac.il>